Задание 12.2-12.3
